Revert 213589 to reland 213396, which re-lands 212635 - Track tab node IDs in SyncedSessionTracker.
Only difference to 213396 is that this also includes: https://codereview.chromium.org/20239002, removing a DCHECK from original cl that landed as 212635.

TBR=pkasting@chromium.org

BUG=98892, 80194, 263999

Review URL: https://chromiumcodereview.appspot.com/20529002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@213872 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5eec4c8caec8a65d3471337a8a8401950f8c0250

+++ chrome/browser/sync/glue/tab_node_pool_unittest.cc
@@ -12,7 +12,7 @@ class SyncTabNodePoolTest : public testing::Test {
  protected:
   SyncTabNodePoolTest() : pool_(NULL) { pool_.SetMachineTag("tag"); }
 
-  size_t GetMaxUsedTabNodeId() const { return pool_.max_used_tab_node_id_; }
+  int GetMaxUsedTabNodeId() const { return pool_.max_used_tab_node_id_; }
 
   TabNodePool pool_;
 };
@@ -24,24 +24,24 @@ TEST_F(SyncTabNodePoolTest, TabNodeIdIncreases) {
   SessionID session_id;
   session_id.set_id(1);
   pool_.AddTabNode(4, session_id, 10);
-  EXPECT_EQ(10u, GetMaxUsedTabNodeId());
+  EXPECT_EQ(10, GetMaxUsedTabNodeId());
   session_id.set_id(2);
   pool_.AddTabNode(5, session_id, 1);
-  EXPECT_EQ(10u, GetMaxUsedTabNodeId());
+  EXPECT_EQ(10, GetMaxUsedTabNodeId());
   session_id.set_id(3);
   pool_.AddTabNode(6, session_id, 1000);
-  EXPECT_EQ(1000u, GetMaxUsedTabNodeId());
+  EXPECT_EQ(1000, GetMaxUsedTabNodeId());
   pool_.ReassociateTabNode(6, 500);
   // Freeing a tab node does not change max_used_tab_node_id_.
   pool_.FreeTabNode(6);
   pool_.FreeUnassociatedTabNodes();
-  EXPECT_EQ(1000u, GetMaxUsedTabNodeId());
+  EXPECT_EQ(1000, GetMaxUsedTabNodeId());
   for (int i = 0; i < 3; ++i) {
     pool_.AssociateTabNode(pool_.GetFreeTabNode(), i + 1);
-    EXPECT_EQ(1000u, GetMaxUsedTabNodeId());
+    EXPECT_EQ(1000, GetMaxUsedTabNodeId());
   }
 
-  EXPECT_EQ(1000u, GetMaxUsedTabNodeId());
+  EXPECT_EQ(1000, GetMaxUsedTabNodeId());
 }
 
 TEST_F(SyncTabNodePoolTest, OldTabNodesAddAndRemove) {
